+++ app/src/main/java/ru/geekbrains/WeatherApi/Clouds.java
@@ -1,4 +1,4 @@
-package ru.geekbrains;
+package ru.geekbrains.WeatherApi;
 
 public class Clouds {
     private int all;
