+++ app/src/main/java/ru/geekbrains/WeatherApi/WeatherRequest.java
@@ -1,4 +1,4 @@
-package ru.geekbrains;
+package ru.geekbrains.WeatherApi;
 
 public class WeatherRequest {
     private Coord coord;
